Creation of the entities Phone and Customer. Creation of a fixture to create fake Data.
