+++ composer.json
@@ -26,7 +26,8 @@
         "symfony/yaml": "5.1.*"
     },
     "require-dev": {
-        "doctrine/doctrine-fixtures-bundle": "^3.3"
+        "doctrine/doctrine-fixtures-bundle": "^3.3",
+        "fzaninotto/faker": "^1.9"
     },
     "config": {
         "optimize-autoloader": true,
