+++ symfony.lock
@@ -96,6 +96,9 @@
     "doctrine/sql-formatter": {
         "version": "1.1.1"
     },
+    "fzaninotto/faker": {
+        "version": "v1.9.1"
+    },
     "jms/metadata": {
         "version": "2.3.0"
     },
